requirements.txt
# PostgreSQL database connector
psycopg2-binary>=2.9.5

# Data manipulation and visualization
pandas>=1.3.0
matplotlib>=3.5.0

# Basic requirements
numpy>=1.20.0
python-dotenv>=0.19.0
requests>=2.26.0
ttkthemes>=3.2.0
pillow>=8.3.0
tkinter>=8.6 